feat: redondeo HALF-UP, compresion GZIP, fix canonicalizacion XML-DSig
- redondeo.py: redondear_monto() con ROUND_HALF_UP (2 decimales), segun
  algoritmo del SIN; aplicado en xml_builder.py a montos/precios/subtotales.
- paquetes.py: comprimir_gzip/descomprimir_gzip para "Recepcion Paquete
  Factura" (contingencia/masiva).
- signing.py: corrige bug latente (cert=[certificate], no objeto suelto -
  signxml espera cert_chain) y usa c14n_algorithm=CANONICAL_XML_1_0 para
  que <SignedInfo>/<CanonicalizationMethod> coincida con el ejemplo Java
  oficial del SIN (antes signxml usaba C14N 1.1 por defecto).
- models/facturacion.py, models/billing.py: corrige type hints de montos
  de Mapped[float] a Mapped[Decimal] (las columnas Numeric ya devuelven
  Decimal en runtime; el hint float era inconsistente y riesgoso).

+++ app/models/billing.py
@@ -1,6 +1,7 @@
 import enum
 import uuid
 from datetime import date, datetime
+from decimal import Decimal
 
 from sqlalchemy import Date, DateTime, Enum, ForeignKey, Numeric, String, func
 from sqlalchemy.dialects.postgresql import UUID
@@ -22,8 +23,8 @@ class Plan(UUIDPKMixin, TimestampMixin, Base):
     __tablename__ = "plans"
 
     nombre: Mapped[str] = mapped_column(String(100), unique=True)
-    precio_mensual: Mapped[float] = mapped_column(Numeric(10, 2))
-    precio_por_factura: Mapped[float] = mapped_column(Numeric(10, 4))
+    precio_mensual: Mapped[Decimal] = mapped_column(Numeric(10, 2))
+    precio_por_factura: Mapped[Decimal] = mapped_column(Numeric(10, 4))
     facturas_incluidas: Mapped[int] = mapped_column(default=0)
 
     suscripciones: Mapped[list["Subscription"]] = relationship(back_populates="plan")
@@ -51,6 +52,6 @@ class UsageRecord(UUIDPKMixin, Base):
     tenant_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("tenants.id"), index=True)
     factura_id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), ForeignKey("facturas.id"), unique=True)
     periodo: Mapped[str] = mapped_column(String(7))  # formato YYYY-MM
-    monto_cobrado: Mapped[float] = mapped_column(Numeric(10, 4))
+    monto_cobrado: Mapped[Decimal] = mapped_column(Numeric(10, 4))
     facturado: Mapped[bool] = mapped_column(default=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
